backend/requirements.txt
azure-cognitiveservices-vision-computervision==0.9.1
fastapi==0.115.8
matplotlib==3.10.0
msrest==0.7.1
Pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn==0.34.0
python-multipart==0.0.20
google-generativeai==0.8.4